refactor: move function to /safe export
